requirements.txt
ipython==8.29.0
langgraph==0.2.45
langchain-community==0.3.5
langchain-anthropic==0.2.4
tavily-python==0.5.0
pandas==2.2.3
python-dotenv==1.0.1
openai==1.54.2